Replace unsigned int with size_t where value indicates a count.

+++ include/sp_tree.h
@@ -42,19 +42,19 @@ typedef struct sp_tree
 
 sp_tree*	sp_tree_new(dict_compare_func cmp_func, dict_delete_func del_func);
 dict*		sp_dict_new(dict_compare_func cmp_func, dict_delete_func del_func);
-unsigned	sp_tree_free(sp_tree *tree);
+size_t		sp_tree_free(sp_tree *tree);
 
 int			sp_tree_insert(sp_tree *tree, void *key, void *datum,
 						   int overwrite);
 int			sp_tree_probe(sp_tree *tree, void *key, void **datum);
 void*		sp_tree_search(sp_tree *tree, const void *key);
 int			sp_tree_remove(sp_tree *tree, const void *key);
-unsigned	sp_tree_clear(sp_tree *tree);
-unsigned	sp_tree_traverse(sp_tree *tree, dict_visit_func visit);
-unsigned	sp_tree_count(const sp_tree *tree);
-unsigned	sp_tree_height(const sp_tree *tree);
-unsigned	sp_tree_mheight(const sp_tree *tree);
-unsigned	sp_tree_pathlen(const sp_tree *tree);
+size_t		sp_tree_clear(sp_tree *tree);
+size_t		sp_tree_traverse(sp_tree *tree, dict_visit_func visit);
+size_t		sp_tree_count(const sp_tree *tree);
+size_t		sp_tree_height(const sp_tree *tree);
+size_t		sp_tree_mheight(const sp_tree *tree);
+uint64_t	sp_tree_pathlen(const sp_tree *tree);
 const void*	sp_tree_min(const sp_tree *tree);
 const void*	sp_tree_max(const sp_tree *tree);
 
@@ -69,8 +69,8 @@ int			sp_itor_valid(const sp_itor *itor);
 void		sp_itor_invalidate(sp_itor *itor);
 int			sp_itor_next(sp_itor *itor);
 int			sp_itor_prev(sp_itor *itor);
-int			sp_itor_nextn(sp_itor *itor, unsigned count);
-int			sp_itor_prevn(sp_itor *itor, unsigned count);
+int			sp_itor_nextn(sp_itor *itor, size_t count);
+int			sp_itor_prevn(sp_itor *itor, size_t count);
 int			sp_itor_first(sp_itor *itor);
 int			sp_itor_last(sp_itor *itor);
 int			sp_itor_search(sp_itor *itor, const void *key);
